Add Koin support and add some teste with injection

+++ src/main/kotlin/br/com/ktortest/dao/AddressTable.kt
@@ -5,6 +5,6 @@ import org.jetbrains.exposed.dao.LongIdTable
 object AddressTable : LongIdTable() {
     val street = varchar("street", 100)
     val zipCode = varchar("zip_code", 25)
-    val number = integer("number")
+    val number = long("number")
     val userId = optReference("fk_user", UserTable)
 }
